+++ src/item/item.module.ts
@@ -2,7 +2,6 @@ import { Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
 import { CqrsModule } from '@nestjs/cqrs';
 
-import { ItemService } from './services';
 import { ApplicationModule } from './application/application.module';
 import { DomainModule } from './domain/domain.module';
 import { InfrastructureModule } from './infrastructure/infrastructure.module';
@@ -16,7 +15,6 @@ import { InfrastructureModule } from './infrastructure/infrastructure.module';
   ],
   providers: [
     ConfigService,
-    ItemService,
   ],
 })
 export class ItemModule {}
